+++ second-semester/Filesystem/src/main/java/ua/knu/elements/manipulator/Manipulator.java
@@ -1,6 +1,6 @@
 package ua.knu.elements.manipulator;
 
 public interface Manipulator {
-    int readInt(byte[] data, int pos);
-    void writeInt(byte[] data, int pos, int value);
+    long readInt(byte[] data, int pos);
+    void writeInt(byte[] data, int pos, long value);
 }
